+++ lib/constants/app_info.dart
@@ -1,8 +1,8 @@
 /// App version and info constants
 /// Update version here only - all other files will reference this
 class AppInfo {
-  static const String version = '2.0.7-preview2';
-  static const String buildNumber = '37';
+  static const String version = '2.1.0-preview';
+  static const String buildNumber = '39';
   static const String fullVersion = '$version+$buildNumber';